# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for Karaoke web app.",
        "title": "Karaoke API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/favorites": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "List favorites",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by channel",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.FavoriteRes"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorRes"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "Add a favorite",
                "parameters": [
                    {
                        "description": "Favorite data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FavoriteReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.FavoriteRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorRes"
                        }
                    }
                }
            }
        },
        "/favorites/{videoId}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Delete favorite by videoId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YouTube Video ID",
                        "name": "videoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorRes"
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search karaoke songs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.KaraokeRes"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorRes"
                        }
                    }
                }
            }
        },
        "/stream/{videoId}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Stream"
                ],
                "summary": "Stream video by videoId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YouTube Video ID",
                        "name": "videoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorRes": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
        "models.FavoriteReq": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "videoId": {
                    "type": "string"
                }
            }
        },
        "models.FavoriteRes": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "videoId": {
                    "type": "string"
                }
            }
        },
        "models.KaraokeRes": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "videoId": {
                    "type": "string"
                }
            }
        }
    }
}